Data.Type.List.Edit

Description

Witnesses regarding single-item edits of lists.

Simple edits

data Insert :: [k] -> [k] -> k -> Type where

An Insert as bs x is a witness that you can insert x into some position in list as to produce list bs.

Some examples:

InsZ                   :: Insert '[1,2,3] '[4,1,2,3] 4
InsS InsZ              :: Insert '[1,2,3] '[1,4,2,3] 4
InsS (InsS InsZ)       :: Insert '[1,2,3] '[1,2,4,3] 4
InsS (InsS (InsS InsZ) :: Insert '[1,2,3] '[1,2,3,4] 4

Constructors

InsZ :: Insert as (x ': as) x

InsS :: Insert as bs x -> Insert (a ': as) (a ': bs) x

data Delete :: [k] -> [k] -> k -> Type where

A Delete as bs x is a witness that you can delete item x from as to produce the list bs. It is essentially Insert flipped.

Some examples:

DelZ             :: Delete '[1,2,3] '[2,3] 1
DelS DelZ        :: Delete '[1,2,3] '[2,3] 2
DelS (DelS DelZ) :: Delete '[1,2,3] '[1,2] 3

Constructors

DelZ :: Delete (x ': as) as x

DelS :: Delete as bs x -> Delete (a ': as) (a ': bs) x

insToDel :: Insert as bs x -> Delete bs as x

delToIns :: Delete as bs x -> Insert bs as x

Flip a deletion

data Substitute :: [k] -> [k] -> k -> k -> Type where

A Substitute as bs x y is a witness that you can replace item x in as with item y to produce bs.

Some examples:

SubZ             :: Substitute '[1,2,3] '[4,2,3] 1 4
SubS SubZ        :: Substitute '[1,2,3] '[1,4,3] 2 4
SubS (SubS SubZ) :: Substitute '[1,2,3] '[1,2,4] 3 4

Constructors

SubZ :: Substitute (x ': as) (y ': as) x y

SubS :: Substitute as bs x y -> Substitute (c ': as) (c ': bs) x y

flipSub :: Substitute as bs x y -> Substitute bs as y x

Flip a substitution

Compound edits

data Edit :: [k] -> [k] -> Type where

An Edit as bs is an edit script transforming as into bs through successive insertions, deletions, and substitutions.

Constructors

ENil :: Edit as as

EIns :: Insert bs cs x -> Edit as bs -> Edit as cs

EDel :: Delete bs cs x -> Edit as bs -> Edit as cs

ESub :: Substitute bs cs x y -> Edit as bs -> Edit as cs

flipEdit :: Edit as bs -> Edit bs as

Reverse an Edit script. O(n^2).

Product

insertProd :: Insert as bs x -> f x -> Prod f as -> Prod f bs

Insert a value into a Prod, at a position indicated by the Insert.

deleteProd :: Delete as bs x -> Prod f as -> Prod f bs

Delete a value in a Prod, at a position indicated by the Delete.

deleteGetProd :: Delete as bs x -> Prod f as -> (f x, Prod f bs)

Retrieve and delete a value in a Prod, at a position indicated by the Delete.

lensProd :: forall as bs x y g f. Functor f => Substitute as bs x y -> (g x -> f (g y)) -> Prod g as -> f (Prod g bs)

A type-changing lens into a value in a Prod, given a Substitute indicating which value. This lens can change the type of the value, unlike lensProd'.

Read this type signature as:

lensProd
    :: Substitute as bs x y
    -> Lens (Prod g as) (Prod g bs) (g x) (g y)

substituteProd :: Substitute as bs x y -> (f x -> f y) -> Prod f as -> Prod f bs

Substitute a value in a Prod at a given position, indicated by the Substitute. This is essentially a specialized version of lensProd.

Index

shiftIndex :: Insert as bs x -> Index as y -> Index bs y

If you add an item to as to create bs, you also need to shift an Index as y to Index bs y. This shifts the Index in as to become an Index in bs, but makes sure that the index points to the same original value.

data Unshifted :: [k] -> k -> k -> Type where

Used as the return type of unshiftIndex. An Unshifted bs x y is like a Maybe (Index bs y), except the Nothing case witnesses that x ~ y.

Constructors

GotDeleted :: Unshifted bs x x

NotDeleted :: Index bs y -> Unshifted bs x y

unshiftIndex :: Delete as bs x -> Index as y -> Unshifted bs x y

If you delete an item in as to create bs, you also need to move Index as y into Index bs y. This transforms the Index in as to become an Index in bs, making sure the index points to the same original value.

However, there is a possibility that the deleted item is the item that the index was originally pointing to. If this is the case, this function returns GotDeleted, a witness that x ~ y. Otherwise, it returns NotDeleted with the unshifted index.

unshiftIndex_ :: Delete as bs x -> Index as y -> Maybe (Index bs y)

A version of unshiftIndex returning a simple Maybe. This can be used if you don't care about witnessing that x ~ y in the case that the index is the item that is deleted.

data Reshifted :: [k] -> k -> k -> k -> Type where

Used as the return type of reshiftIndex. An Reshifted bs x y z is like an Either (Index bs y) (Index bs z), except the Left case witnesses that x ~ z.

Constructors

GotSubbed :: Index bs y -> Reshifted bs z y z

NotSubbed :: Index bs z -> Reshifted bs x y z

reshiftIndex :: Substitute as bs x y -> Index as z -> Reshifted bs x y z

If you substitute an item in as to create bs, you also need to reshift Index as z into Index bs z. This reshifts the Index in as to become an Index in bs, making sure the index points to the same original value.

However, there is a possibility that the substituted item is the item that the index was originally pointing to. If this is the case, this function returns GotSubbed, a witness that x ~ z. Otherwise, it returns NotSubbed. Both contain the updated index.

reshiftIndex_ :: Substitute as bs x y -> Index as z -> Either (Index bs y) (Index bs z)

A version of reshiftIndex returning a simple Either. This can be the case if you don't care about witnessing x ~ z in the case that the index is the item that was substituted.