Data.Type.List.Prod

Documentation

data Prod :: (k -> Type) -> [k] -> Type where

This shows up in many places in the Haskell ecosystem: it's vinyl, or a heterogeneous list parameterized by a functor.

A value of type

Prod f '[a,b,c]

contains an f a, and f b, and an f c.

Constructors

Ø :: Prod f '[]

(:<) :: f a -> Prod f as -> Prod f (a ': as) infixr 5

pNil :: Prod f '[]

Alias for Ø

indexProd :: Index as x -> Prod f as -> f x

Retrieve an element in a Prod at a given index.

lensProd' :: forall as x g f. Functor f => Index as x -> (g x -> f (g x)) -> Prod g as -> f (Prod g as)

A type-preserving lens into a value in a Prod, given a Index indicating which value. This lens cannot change the type of the value, unlike "Data.Type.List.Edit.lensProd'.

Read this type signature as:

lensProd
    :: Index as x
    -> Lens' (Prod g as) (g x)

headProd :: Prod f (a ': as) -> f a

tailProd :: Prod f (a ': as) -> Prod f as