Data.Type.List.Sublist

Documentation

data Prefix :: [k] -> [k] -> Type where

A Prefix as bs witnesses that as is a prefix of bs.

Some examples:

PreZ                    :: Prefix '[]      '[1,2,3]
PreS PreZ               :: Prefix '[1]     '[1,2,3]
PreS (PreS PreZ)        :: Prefix '[1,2]   '[1,2,3]
PreS (PreS (PreS PreZ)) :: Prefix '[1,2,3] '[1,2,3]

Rule of thumb for construction: the number of PreS is the number of items in the prefix.

Constructors

PreZ :: Prefix '[] as

PreS :: Prefix as bs -> Prefix (a ': as) (a ': bs)

data Suffix :: [k] -> [k] -> Type where

A Suffix as bs witnesses that as is a suffix of bs.

Some examples:

SufZ                      :: Suffix '[1,2,3] '[1,2,3]
SufS SufZ                 :: Suffix   '[2,3] '[1,2,3]
SufS (SufS SufZ)          :: Suffix     '[3] '[1,2,3]
SufS (SufS (SufS (SufZ))) :: Suffix      '[] '[1,2,3]

Rule of thumb for construction: the number of SufS is the number of items to "drop" before getting the suffix.

Constructors

SufZ :: Suffix as as

SufS :: Suffix as bs -> Suffix as (b ': bs)

takeProd :: Prefix as bs -> Prod f bs -> Prod f as

Take items from a Prod corresponding to a given Prefix.

dropProd :: Suffix as bs -> Prod f bs -> Prod f as

Drop items from a Prod corresponding to a given Suffix.