Data.Type.List.Prod

Description

Parameterized heterogeneous list.

Documentation

data Prod :: (k -> Type) -> [k] -> Type where

This shows up in many places in the Haskell ecosystem: it's vinyl, or a heterogeneous list parameterized by a functor.

A value of type

Prod f '[a,b,c]

contains an f a, and f b, and an f c.

While this exists in my places, it's reproduced here to to provide a canonical implementation.

Constructors

Ø :: Prod f '[]

(:<) :: f a -> Prod f as -> Prod f (a ': as) infixr 5

pNil :: Prod f '[]

Alias for Ø

headProd :: Prod f (a ': as) -> f a

(Total) head of a Prod.

tailProd :: Prod f (a ': as) -> Prod f as

(total) tail of a Prod.

lensProd' :: forall as x g f. Functor f => Index as x -> (g x -> f (g x)) -> Prod g as -> f (Prod g as)

A type-preserving lens into a value in a Prod, given a Index indicating which value. This lens cannot change the type of the value, unlike "Data.Type.List.Edit.lensProd'.

Read this type signature as:

lensProd
    :: Index as x
    -> Lens' (Prod g as) (g x)

indexProd :: Index as x -> Prod f as -> f x

Retrieve an element in a Prod at a given index.

Traversals

mapProd :: forall f g as. (forall x. f x -> g x) -> Prod f as -> Prod g as

A RankN map over items in a Prod.

traverseProd :: forall f g h as. Applicative h => (forall x. f x -> h (g x)) -> Prod f as -> h (Prod g as)

A RankN traversal over items in a Prod.

foldMapProd :: forall f as m. Monoid m => (forall x. f x -> m) -> Prod f as -> m

RankN fold over items in a Prod.

zipProd :: Prod f as -> Prod g as -> Prod (f :*: g) as

Zip together two Prod.

With Singletons

prodSing :: Prod Sing as -> Sing as

Convert a Prod of Sing elements into a Sing of the list of elements.

singProd :: Sing as -> Prod Sing as

Convert a Sing of a list of elements into a Prod of Sing elements.